Made rough implementation of Graph, Set and SAT
Now I need to test these in conjunction with one another. Set could
perhaps be improved upon, but I will not beautify any code before this
works. It also does not make sense to optimize insane amounts while I do
not know the problem domain
But rough
Implicit design decision:
Bimap of description <-> atoms is a structure OUTSIDE the graph, because
it has to be able to operate on modifications of the graph.

The same goes for bimap of description <-> states.

It does not have to be this way, a more sensible solution could be to
make sure that each bimap is bound to the graph, this way state_ids
could just be their position in the array, and atoms could likewise be
contigious integers.

The main idea is that the input of atoms will never be integers,
(because they are pretty unintelligle) but always a description of what
the states mean and what the atoms mean.

But this optimization can be made if necessary in next iteration.

Graph: A simple graph specifying the state_ids and the nodes

Set: a sorted list of u64

SAT solver: solves the problem recursively, caller is responsible for
freeing the memory allocated by calling the SAT solver.