Began imp of hanabi game, to be used by AI
I will try to make a game that can be used for the simulation
I try to have a natural seperation between what is known by a player
and what is known by the internal game state, so that a player cannot
accidentally "cheat" in what it knows.
Right now I try not to code rules in the agents themselves.
But it could be interesting further work to look into the work of baltag
and see whether it allows for some interesting solutions of hanabi.
But right now there will only be time for a more naive approach where I
use principle of modal logic, but cannot make use of it in its entirety.