Implementated combinatorics in zig
My method was to implement it initially in python, mainly because I was
not sure how it would be implemented exactly, but once I did it and it
worked, I moved it into zig.
Furthermore I have written some notes about the representation of the
hand and deck etc. And it can be done quite uniformly and compactly.
Which is great. Using the deck pool structure.
That is making use of the fact that there are no more than 3 of each
card, then I can make an array [25]u2, and just describe how many of
each card there is, this is compact and straightforward. Could it be
made more compact as an integer? I don't think so.
Furthermore it seems likely that the number of combinations a player
will have to consider is about 10000, which also puts limits on how the
representation must be made in terms of size.

The combinatorics method is from https://math.stackexchange.com/questions/1902850/efficient-algorithm-to-find-all-unique-combinations-of-set-given-duplicates
using the contingency table method.
And is pretty efficient (I could try to benchmark this method against
generating everything for the sake of the report)
But anyways its great.

Up next: Write some things in the report, set up the structure and
write about design decisions.