Combinatorial helpers are in place
Now the combinatorial helpers are somewhat done.
Things can be done in order to optimize for space, i.e. using packed
arrays for the [25]u2, since I forgot that "a byte" is the lowest
possible addressable space, so it does not pack them by default.
Therefore [25]u2 is not 50 bits, but it is actually 25 bytes!. Lots of
wasted space.

Using memory allocators combined with trying to make a huge array might
pose a problem, but we will wait and see whether the arrays seem
coherent in regards to the "buckets" of the players. I.e. I expect a
fixed card of POV player will result in another player having all its
possibilities close in memory relative to that fixed card. We will see
:)

I am working all I can on doing my MVP, but time is short :(