Various,Implemented layer 2 removal based on hints
Strategy
========
I also did some other things,
I wrote a strategy in todo based on the personal knowledge.
Lots of the strategy is based on Cox et al article on near perfect
hanabi strategies.
I am close to having a agent that can decide whether it got some card
that is playable or not.
Layer 2 implementation is based on removal, where layer 1 is based
solely on generation.

Allocator
=========
I don't need a general purpose
allocator, I can just assume some small number of bytes, because we are
often talking about hands. This leads to simpler, less error prone code,
because it just respects the scope lifetime. While having the benefits
of arraylists

Reflection about the lack of use of metaknowledge
=================================================
If there is something I want to make different was to make more use
of the knowledge about the other players, right now it is only used for
deciding whether a fixed scenario can exist. i.e. if a fixed scenario
leads some other player to have no imagined hand, then that fixed
scenario cannot be the case.

A very good use case for this would be to augment the information
strategy deviced by Cox et al, but giving it better estimates based on
the knowledge system deviced here.