Close to finishing up the agents
I need to implement the generation method and then the strategy.
Then I have to do various tests

The generation basically assumes that each player is init and then
deinit, this makes sense since most of the players content is the
kripkestructure. So I am just gonna do that.

I might have to optimize for space before I can run a fully fledged
simulation. But I will wait and see.

I think there is potential that this method can be used for a minihanabi
version. Many of the principle are identical, but I will initially
implement a full hanabi version and then see if there is some patterns
that emerges that I can take advantage of.