Strategy still ongoing
I wrote in the todo how one might make use of the knowledge base in
order to give better hints, but right now in order to simplify
implementation and be done in proper time I am choosing any random
"unknown" card color or value, and hinting that.
This will give information but probably not a lot.